Fix windows warnings, turn on werror in build 2017

+++ src/windtoolkit.cpp
@@ -205,8 +205,8 @@ wxString WindToolkitDialog::GetYear()
 wxArrayString WindToolkitDialog::GetHubHeights()
 {
 	wxArrayString hh;
-	for (int i = 0; i < lstHubheights->GetCount(); i++)
-		if (lstHubheights->IsSelected(i))
+	for (size_t i = 0; i < lstHubheights->GetCount(); i++)
+		if (lstHubheights->IsSelected((int)i))
 			hh.Add(lstHubheights->GetString(i));
 	return hh;
 }
